Odin port: Julia-exact RNG/exp, coupled parity, ensemble + per-parcel observable
--rng julia replays Julia 1.12 MersenneTwister bit-for-bit (pure-Odin
dSFMT19937 + 1002/501 caches + NDL ranges; 4320/4320 values verified).
jexp_f64 ports Julia's inline base/special/exp.jl (4015/4015 bit-exact at
-o:none and -o:speed); libm differs by 1 ulp on ~10% of inputs, which
broke N=20 field parity until ported.

f64 fields with exact-order scalar path in Julia mode; --trace diffs
identical vs Julia including the coupled PDE (N=16 seeds 42/7/123, N=20
100 MCS). serial_seed42.csv fixture reproduces 7/7 plus membrane m.

256-seed N=40/p6/400 run: 92.6% input-ordered at MCS 100 decaying to
74.2% at 400; site vs parcel averages disagree on 3/1024 pairs (seed 43
MCS 100 flips sign between them). Seeds 42-57 subset reproduces the
Julia table; MCS-200 CS-top is 15/16 (seed 54, CN>CS clearly).

Exactness-preserving speedups (~1.5x: lazy NDL threshold, force_inline
accessors, hoisted radial loops, -no-bounds-check production builds).
Threads measured bit-exact but net-zero, reverted.

odin test tests/ 10/10. Ensemble CSV (1.2MB, regenerable in ~40s) and
render media stay out of the tree via .gitignore.

+++ odin/tests/coupled_test.odin
@@ -22,7 +22,7 @@ run_coupled :: proc(p: cpm.CPM_Params, seed: u64, n_mcs: int) -> (cpm.Sim, cpm.R
 		}
 		cpm.radiolysis_step(&rd, rd.params.dt_rd)
 		cpm.radial_to_3d(&s, &rd)
-		cpm.update_fields(&s, f32(rd.m))
+		cpm.update_fields(&s, rd.m)
 		cpm.update_centers_of_mass(&s)
 	}
 	return s, rd
